+++ client/src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import InitContainer from './../containers/InitContainer.js';
-import Reviews from './Reviews';
+import Reviews from './reviews/Reviews';
 
 class App extends React.Component {
